adding CoL and updating scripts
CoL: through V6C2

had to update download-chapters to cleanup title better, and properly
replace a <b> with <strong> under the hood (forgot the .string...)
Also started stripping specifying font size and type, it just made
things look really really nasty in the epub.

in make-epub, added in ability to not process 'chapters' tag at all,
CoL has some url's that are randomly in a different format, so they
are explicitly specified with 'chapters' and would then be doubly included
by make-epub, which would be an error. so we can now skip chapters all-together
and we also won't include a file if its already been added.